Rename zlib.c to zlib_wrapper.c to prevent confusion
